Modified keymap, set up for EE_PROM side mastery
Want the ability to set either hand to master in the future, so setting
up the EE_PROM to notify which hand is left and which is right. This
also reconfigures the right-hand thumb-cluster alt to be a reset button,
much like the same position on the left-hand was. I realized I wouldn't
be able to fully reprogram the keyboard without a reset on both halves.

Might make more sense to hide the reset in a layer later.