Added an experimental folder for the other hand
This directory is an experimental key layout that I'm using for the
right-hand-primary configuration. I plan to use this for things that I'm
trying out and am not sure if I like yet or not.

Specifically in this key layout I moved the enter and the backspace keys
down to the lower section of the thumb cluster; I put another space
button in the place of enter; The equals/+ button got moved up to where
backspace used to be, and end got moved to replace PGDN, which has been
removed and is only represented by keys in the left-hand layer.